Update to stable release 28 and replace gtk-3.0 with gtk-3.20
Previous release r18 used folder gtk-3.0, while the new r28 splitted the
support in gtk-3.0 and gtk-3.20. I could not find (yet) a way to simply
use the new folder location, so I added a script to replace the content
of folder gtk-3.0 with the content in gtk-3.20.

Unless flatpak theming requires the folder to be named gtk-3.0, this
would be a temporary workaround